refactor(slacksvc): move run-health thread seam to notifier_health.go (PRD #1026 M1)
